+++ CHANGELOG.md
@@ -25,6 +25,8 @@ A feature release on top of `1.4.0`, backwards compatible.
 
 ### Fixes
 
+- `just-import-git` now correctly inherits pragmas for imported
+  file-type repositories during description rewrite.
 - `just-mr` repository garbage collection now properly removes
   no longer needed directories.
 - The "generic" rule now properly detects staging conflicts.